Wipe minibuf on hide to prevent retained buffer with incorrect state.
By default, when the minibuf explorer window is hidden the buffer
becomes unlisted.  During this process it loses all of its config,
including the buftype.  This makes it susceptible to buffer rename
during cd: commands as will occur if files are passed through the 'vim
--remote' command line interface.  As a result, window splits do not
occur correctly and the minibuf window has incorrect contents.